backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
python-multipart==0.0.20
docling
PyYAML==6.0.2
